+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2] - 2023-07-31
+
+### Added
+
+- Added /clear command to clear the whole queue
+- Added /remove command to remove a song from the queue with an index
+- Added the option to choose to stream the audio or download the audio (Download mode works more stable but will take up more space)
+
+### Fixed
+
+- Fixed a bug where the bot will fail to function when sending multiple commands together
+- Fixed a bug where the bot will fail to function when the queue is empty and adding more songs
+
 ## [1.1] - 2023-07-31
 
 ### Added